refactor(auth): use current-user helpers in shared P0 libs
